requirements.txt
mcp[cli]>=1.20.0
httpx>=0.27.0
pyyaml>=6.0
python-dotenv>=1.0.0
starlette>=0.37.0
uvicorn>=0.29.0
requests>=2.31.0
